agent_template_54321/requirements.txt
python-dotenv
langchain_openai
langgraph
langchain_community
langchain_core